// src/app/private/cliente/models/legal-customer-response.ts
export class LegalCustomerResponse {
  id: number = 0;
  ruc: string = "";
  razonSocial: string = "";
  nombreComercial: string = "";
  tipo: string = "";
  estado: string = "";
  condicion: string = "";
  direccion: string = "";
  departamento: string = "";
  provincia: string = "";
  distrito: string = "";
  ubigeo: string = "";
  capital: string = "";
  idPersonaTipoDocumento: number = 0;
}